requirements.txt
pyyaml==6.0.2
simple-term-menu==1.6.4
colorama==0.4.6
